% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_cyl}
\alias{filter_cyl}
\title{Filter mtcars by cylinder}
\usage{
filter_cyl(.mtcars, .cyl)
}
\arguments{
\item{.mtcars}{Filtered or unfiltered mtcars data}

\item{.cyl}{A numeric cylinder value of 4, 6, or 8}
}
\value{
mtcars data filtered by cylinder
}
\description{
Filter mtcars by cylinder
}
