% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_max_mpg}
\alias{filter_max_mpg}
\title{Filter mtcars by max mpg}
\usage{
filter_max_mpg(.mtcars, .mpg)
}
\arguments{
\item{.mtcars}{Filtered or unfiltered mtcars data}

\item{.mpg}{A numeric value of miles per gallon}
}
\value{
mtcars data filtered with miles per gallon less than '.mpg'
}
\description{
Filter mtcars by max mpg
}
