% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc.shinymeta-package.R
\docType{package}
\name{poc.shinymeta-package}
\alias{poc.shinymeta}
\alias{poc.shinymeta-package}
\title{poc.shinymeta: PoC shinymeta}
\description{
A proof of concept for the shinymeta package.
}
\author{
\strong{Maintainer}: Andrew McNeil \email{andrew.mcneil@sanofi.com}

}
\keyword{internal}
