% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_filters.R
\name{select_cols}
\alias{select_cols}
\title{Select mtcars columns}
\usage{
select_cols(.mtcars, .cols)
}
\arguments{
\item{.mtcars}{Filtered or unfiltered mtcars data}

\item{.cols}{A character vector of column names to select}
}
\value{
mtcars data containing a subset of columns
}
\description{
Select mtcars columns
}
